+++ R/batchtools_lsf.R
@@ -59,7 +59,7 @@ BatchtoolsLsfFutureBackend <- function(...) {
 #' * <https://en.wikipedia.org/wiki/IBM_Spectrum_LSF>
 #'
 #' @export
-batchtools_lsf <- function(..., template = "lsf", scheduler.latency = 1.0, fs.latency = 65.0, resources = list(), workers = getOption("future.batchtools.workers", default = 100L)) {
+batchtools_lsf <- function(..., template = "lsf", scheduler.latency = 1.0, fs.latency = 65.0, resources = list(), delete = "on-success", workers = getOption("future.batchtools.workers", default = 100L)) {
  stop("INTERNAL ERROR: The future.batchtools::batchtools_lsf() must never be called directly")
 }
 class(batchtools_lsf) <- c(